server/requirements.txt
aiofiles==0.8.0
anyio==3.6.1
async-timeout==4.0.2
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
Deprecated==1.2.13
ecdsa==0.18.0
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
jmespath==1.0.1
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
tomli==2.0.1
types-requests==2.28.9
types-urllib3==1.26.23
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
wrapt==1.14.1
yapf==0.32.0
